// superset-frontend/plugins/plugin-chart-bulk-loading-card/src/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import { QueryFormData, TimeseriesDataRecord } from '@superset-ui/core';

export interface BulkLoadingCardStylesProps {
  height: number;
  width: number;
}

export interface BulkLoadingCardCustomizeProps {
  // Column Mappings
  workOrderColumn: string;
  materialNameColumn?: string;
  secondaryIdColumn?: string;
  statusColumn?: string;
  destinationTankColumn?: string;
  elapsedTimeColumn?: string;

  // Visual Customizations
  titleText: string;
  showTruckGraphic: boolean;
  showStatusPipeline: boolean;
  activeStatusColor: string;
  cardBgColor: string;
  cardBorderColor: string;
  fontSize: number;
  workOrderFontSize: number;
  timerColor: string;
  tankBadgeColor: string;
}

export interface BulkLoadingCardProps extends BulkLoadingCardStylesProps {
  data: TimeseriesDataRecord[];
  customize: BulkLoadingCardCustomizeProps;
}

export type BulkLoadingCardQueryFormData = QueryFormData &
  BulkLoadingCardStylesProps &
  BulkLoadingCardCustomizeProps;